Combine node files to make mocha tests possible
