+++ README.md
@@ -16,6 +16,7 @@ Currently working methods of nodes are:
  - Get list of known clients. Both earlier and currently connected clients
  - Get signal strength. [Made by request in Issue#1](https://github.com/zinen/node-red-contrib-huawei-router/issues/1)
  - Toggle mobile data connection
+ - Reboot device
 
 ## Known limits
- - The router can only have so many active sessions a session seems to end after 5 minutes of inactivity. The limit in active sessions means you can get a error code back from this router `108003: Already login` meaning too active sessions. This is a limit on the router it self. 
+ - The router can only have so many active sessions and a session seems to end after 5 minutes of inactivity. The limit in active sessions means you can get a error code back from this router `108003: Already login` meaning too many active sessions. This is a limit on the router it self. Wait some time (max 5 minutes) and try again.
