--- huawei-router/huawei-config.html
@@ -1,28 +1,0 @@
-<script type="text/javascript">
-    RED.nodes.registerType('huawei-config',{
-        category: 'config',
-        defaults: {
-            user: { value: "admin", required: true },
-            pass: { value: "", required: true },
-            url: { value: "192.168.1.1", required: true }
-        },
-        label: function() {
-            return this.user+"@"+this.url;
-        }
-    });
-</script>
-
-<script type="text/html" data-template-name="huawei-config">
-    <div class="form-row">
-        <label for="node-config-input-user"><i class="fa fa-user"></i> Username</label>
-        <input type="text" id="node-config-input-user" placeholder="Eg. admin">
-    </div>
-    <div class="form-row">
-        <label for="node-config-input-pass"><i class="fa fa-key"></i> Password</label>
-        <input type="password" id="node-config-input-pass" placeholder="" autocomplete="off">
-    </div>
-    <div class="form-row">
-        <label for="node-config-input-url"><i class="fa fa-globe"></i> URL of router</label>
-        <input type="text" id="node-config-input-url" placeholder="Eg. 192.168.0.45">
-    </div>
-</script>
