--- huawei-router/huawei-lanhosts.html
@@ -1,36 +1,0 @@
-<script type="text/javascript">
-    RED.nodes.registerType('huawei-lanhosts', {
-        category: 'homesmart',
-        color: '#E5766A',
-        defaults: {
-            server: {value:"", type:"huawei-config"}
-        },
-        inputs: 1,
-        outputs: 1,
-        icon: "bridge.svg",
-        label: function () {
-            return this.name || "huawei lanhosts"
-        }
-    })
-</script>
-
-<script type="text/html" data-template-name="huawei-lanhosts">
-    <div class="form-row">
-        <label for="node-input-server"><i class="fa fa-globe"></i> Server</label>
-        <input type="text" id="node-input-server">
-    </div>
-</script>
-
-<script type="text/html" data-help-name="huawei-lanhosts">
-    <p>Returns array of connected devices.</p>
-    <p>Accepts any input to make it run.</p>
-    <h3>Output</h3>
-    <dl class="message-properties">
-        <dt>payload
-            <span class="property-type">array</span>
-        </dt>
-        <dd>
-            Returns array of connected devices 
-        </dd>
-    </dl>
-</script>
