--- huawei-router/huawei-mobiledata.html
@@ -1,75 +1,0 @@
-<script type="text/javascript">
-    RED.nodes.registerType("huawei-mobiledata", {
-        category: 'homesmart',
-        color: '#E5766A',
-        defaults: {
-            server: {value:"", type:"huawei-config"},
-            mode: { value: "off-on"}
-        },
-        inputs: 1,
-        outputs: 1,
-        icon: "bridge.svg",
-        label: function () {
-            return this.name || "huawei mobiledata"
-        }
-    })
-</script>
-
-<script type="text/html" data-template-name="huawei-mobiledata">
-    <div class="form-row">
-        <label for="node-input-server"><i class="fa fa-globe"></i> Server</label>
-        <input type="text" id="node-input-server">
-    </div>
-    <!-- <div class="form-row">
-        <label for="node-input-user"><i class="fa fa-user"></i> Username</label>
-        <input type="text" id="node-input-user" placeholder="Eg. admin">
-    </div>
-    <div class="form-row">
-        <label for="node-input-pass"><i class="fa fa-key"></i> Password</label>
-        <input type="password" id="node-input-pass" placeholder="" autocomplete="off">
-    </div>
-    <div class="form-row">
-        <label for="node-input-url"><i class="fa fa-globe"></i> URL of router</label>
-        <input type="text" id="node-input-url" placeholder="Eg. 192.168.0.45">
-    </div> -->
-    <div class="form-row">
-        <label for="node-input-mode"><i class="fa fa-pencil-square-o"></i> Mode</label>
-        <select value="" id="node-input-mode">
-            <option value="off-on">off-on</option>
-            <option value="on">on</option>
-            <option value="off">off</option>
-            <option value="toggle">toggle</option>
-        </select>
-    </div>
-</script>
-
-<script type="text/html" data-help-name="huawei-mobiledata">
-    <p>Change state of mobile data.</p>
-    <p>Accepts any input to make it run.</p>
-    <h3>Output</h3>
-    <dl class="message-properties">
-        <dt>payload
-            <span class="property-type">string</span>
-        </dt>
-        <dd>
-            Returns OK or throws an error
-        </dd>
-    </dl>
-    <dl class="message-properties">
-        <dt>state
-            <span class="property-type">number</span>
-        </dt>
-        <dd>
-            States wether mobile data is turned on. 1 = ON.
-        </dd>
-    </dl>
-    <h3>Input</h3>
-    <dl class="message-properties">
-        <dt>mode
-            <span class="property-type">string</span>
-        </dt>
-        <dd>
-            Accepted values are 'on', 'off', 'toggle', 'off-on'.
-        </dd>
-    </dl>
-</script>
