--- huawei-router/huawei-signal.html
@@ -1,36 +1,0 @@
-<script type="text/javascript">
-    RED.nodes.registerType('huawei-signal', {
-        category: 'homesmart',
-        color: '#E5766A',
-        defaults: {
-            server: {value:"", type:"huawei-config"}
-        },
-        inputs: 1,
-        outputs: 1,
-        icon: "bridge.svg",
-        label: function () {
-            return this.name || "huawei signal strength"
-        }
-    })
-</script>
-
-<script type="text/html" data-template-name="huawei-signal">
-    <div class="form-row">
-        <label for="node-input-server"><i class="fa fa-globe"></i> Server</label>
-        <input type="text" id="node-input-server">
-    </div>
-</script>
-
-<script type="text/html" data-help-name="huawei-signal">
-    <p>Returns signal strength of data/LTE connection.</p>
-    <p>Accepts any input to make it run.</p>
-    <h3>Output</h3>
-    <dl class="message-properties">
-        <dt>payload
-            <span class="property-type">object</span>
-        </dt>
-        <dd>
-            Returns list of signal data properties. <a href="https://github.com/zinen/node-red-contrib-huawei-router/issues/1">See more info here</a>
-        </dd>
-    </dl>
-</script>
