+++ test/huawei-mobiledata_spec.test.js
@@ -1,21 +1,24 @@
-/* global describe, beforeEach, afterEach, it */
+/* global describe, before, after, afterEach, it */
 /* eslint-disable-next-line no-unused-vars */
 const should = require('should')
 const helper = require('node-red-node-test-helper')
-const testNode = require('../huawei-router/huawei-mobiledata.js')
+const testNode = require('../huawei-router/huawei-router.js')
 require('dotenv').config()
 
 helper.init(require.resolve('node-red'))
 
 describe('huawei-mobiledata Node', function () {
   this.retries(2)
-  beforeEach(function (done) {
+  before(function (done) {
     helper.startServer(done)
   })
 
-  afterEach(function (done) {
-    helper.unload()
+  after(function (done) {
     helper.stopServer(done)
+  })
+
+  afterEach(function () {
+    helper.unload()
   })
 
   it('should be loaded', function (done) {
@@ -28,10 +31,18 @@ describe('huawei-mobiledata Node', function () {
   })
   it('should turn data off-on', function (done) {
     const flow = [
-      { id: 'n1', type: 'huawei-mobiledata', wires: [['n2']], user: process.env.ROUTER_USER, pass: process.env.ROUTER_PASSWORD, url: process.env.ROUTER_URL },
+      {
+        id: 'n0',
+        type: 'huawei-config',
+        user: process.env.ROUTER_USER,
+        pass: process.env.ROUTER_PASSWORD,
+        url: process.env.ROUTER_URL
+      },
+      { id: 'n1', type: 'huawei-mobiledata', wires: [['n2']], server: 'n0' },
       { id: 'n2', type: 'helper' }
     ]
     helper.load(testNode, flow, function () {
+      helper.getNode('n0')
       const n1 = helper.getNode('n1')
       const n2 = helper.getNode('n2')
       n2.on('input', function (msg) {
