+++ test/huawei-reboot_spec.test.js
@@ -1,0 +1,32 @@
+/* global describe, before, after, afterEach, it */
+/* eslint-disable-next-line no-unused-vars */
+const should = require('should')
+const helper = require('node-red-node-test-helper')
+const testNode = require('../huawei-router/huawei-router.js')
+require('dotenv').config()
+
+helper.init(require.resolve('node-red'))
+
+describe('huawei-mobiledata Node', function () {
+  this.retries(2)
+  before(function (done) {
+    helper.startServer(done)
+  })
+
+  after(function (done) {
+    helper.stopServer(done)
+  })
+
+  afterEach(function () {
+    helper.unload()
+  })
+
+  it('should be loaded', function (done) {
+    const flow = [{ id: 'n1', type: 'huawei-reboot', name: 'huawei-reboot' }]
+    helper.load(testNode, flow, function () {
+      const n1 = helper.getNode('n1')
+      n1.should.have.property('name', 'huawei-mobiledata')
+      done()
+    })
+  })
+})
